LPS-163052 Move method, no logical changes
