+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/check/JSPLanguageUtilCheck.java
@@ -14,8 +14,11 @@
 
 package com.liferay.source.formatter.check;
 
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.source.formatter.check.util.JavaSourceUtil;
 
+import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -38,9 +41,58 @@ public class JSPLanguageUtilCheck extends BaseFileCheck {
 			}
 		}
 
+		return _replaceLiferayTag(content);
+	}
+
+	private String _replaceLiferayTag(String content) {
+		Matcher matcher = _javaSourcePattern.matcher(content);
+
+		StringBuffer sb = new StringBuffer();
+
+		while (matcher.find()) {
+			String expression = matcher.group();
+
+			expression = expression.replaceAll("\n\t*", StringPool.BLANK);
+
+			Matcher expMatcher = _languageUtilGetPattern.matcher(expression);
+
+			if (expMatcher.find()) {
+				String javaSource = expMatcher.group(1);
+
+				List<String> parameters = JavaSourceUtil.getParameterList(
+					javaSource);
+
+				if (parameters.size() != 2) {
+					continue;
+				}
+
+				String secondParameter = parameters.get(1);
+
+				if (!secondParameter.startsWith("\"") ||
+					!secondParameter.endsWith("\"") ||
+					(StringUtil.count(secondParameter, "\"") > 2)) {
+
+					continue;
+				}
+
+				matcher.appendReplacement(
+					sb, "<liferay-ui:message key=" + secondParameter + " />");
+			}
+		}
+
+		if (sb.length() > 0) {
+			matcher.appendTail(sb);
+
+			return sb.toString();
+		}
+
 		return content;
 	}
 
+	private static final Pattern _javaSourcePattern = Pattern.compile(
+		"(?<!['\"])<%=((?!%>)[\\s\\S])*%>");
+	private static final Pattern _languageUtilGetPattern = Pattern.compile(
+		"<%= ?(LanguageUtil\\.get\\(request, ((?!\\)).)+\\)) ?%>");
 	private static final Pattern _languageUtilPattern = Pattern.compile(
 		"LanguageUtil\\.get\\(locale,");
